Fixed Issue #212
Changed ReflectiveWithExamples test to have an AndGiven step which
produced the additional redundant AndThen step. Modified
MethodNameStepScanner to return the first match it finds. Not sure why
it would return multiple steps for one method (unless it is to do with
RunStepWithArgs)?  No tests failed so I'm hoping that's not the case.